subscriber_processor/requirements.txt
pymongo==4.5.0
paho-mqtt==1.6.1
requests==2.31.0
numpy==1.22.0
pandas==1.3.5
Flask==2.0.2
Werkzeug==2.0.2